==> analyses/02_main/01_data_preprocessing.Rmd <==
---
title: "Conformity of Norms Replication Preprocessing"
author: "Group 5"
date: "August 27,2021"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE, error = F, message = F, warning = F)
```

```{r libraries, include=FALSE, message=FALSE, warning=FALSE}

library(tidyverse)
library(readr)
library(dplyr)

# global color scheme / non-optimized
project_colors = c("#E69F00", "#56B4E9", "#009E73", "#F0E442", "#0072B2", "#D55E00", "#CC79A7", "#000000")

# setting theme colors globally
scale_colour_discrete <- function(...) {
  scale_colour_manual(..., values = project_colors)
}
scale_fill_discrete <- function(...) {
   scale_fill_manual(..., values = project_colors)
} 

# nicer global knitr options
knitr::opts_chunk$set(warning = FALSE, message = FALSE, 
                      cache = TRUE, fig.align = 'center')

```

### Importing the data

```{r, results='hide'}

df <- read_csv("01_raw_data.csv")

```


### Glimps at the data

```{r}

df %>% 
  head()

```

```{r}

df %>% 
  glimpse()

```


### Unique `submission_id`s with their counts

```{r}

df %>%
  group_by(submission_id) %>% 
  count() %>% 
  unique()

```


### Trials and their count according to `trial_name`

```{r}

df %>% 
  group_by(trial_name) %>% 
  tally()

```

## --- ` For Main Analysis ` ---

### Omitting the unnecessary columns and trials for the analysis

```{r}

df_cleaned <- df %>% 
  select(
    -c(
      QUD,
      RT,
      optionLeft,
      optionRight,
      startDate,
      startTime,
      trial_number
    )
  ) %>% 
  filter(
    trial_name != "choice_of_political_topic" &
      trial_name != "fit_backstory_fake_rating" &
      trial_name != "understanding_check2"
  )

```

### Trials and their count according to `trial_name` after cleaning

```{r}

df_cleaned %>% 
  group_by(trial_name) %>% 
  tally()

```

### Glimps after columns are removed

```{r}

df_cleaned %>% 
  glimpse()

```


### Removing participants who failed the understanding check

```{r}

df_cleaned <- df_cleaned %>% 
  mutate(
    invalid = case_when(
      trial_name == "understanding_check" &
        response == "incorrect" ~ TRUE,
      TRUE ~ FALSE
      ),
    )

invalid_ids <- df_cleaned$submission_id[which(df_cleaned$invalid == TRUE)]

removed_invalid <- sum(df_cleaned$invalid)

for (id in invalid_ids) {
  df_cleaned <- df_cleaned %>% 
    filter(
      submission_id != id
    )
  }


print(
  paste0(
    "We excluded ", removed_invalid, " participant(s) for failing the understanding check."
    )
  )

```


### Excluding participants who were neutral about their chosen social issue 

```{r}

df_cleaned <- df_cleaned %>% 
  mutate(
    neutral = case_when(
      trial_name == "rate_statement" &
        response == 0 ~ TRUE,
      TRUE ~ FALSE
      )
    )

neutral_ids <- df_cleaned$submission_id[which(df_cleaned$neutral == TRUE)]

removed_neutral <- sum(df_cleaned$neutral)

for (id in neutral_ids) {
  df_cleaned <- df_cleaned %>% 
    filter(
      submission_id != id
    )
  }


print(
  paste0(
    "We excluded ", removed_neutral, " participant(s) for being neutral about their chosen social activity."
    )
  )

```


### Dropping temporary made `invalid` and `neutral` columns

```{r}

df_cleaned <- df_cleaned %>% 
  select(
    -c(
      invalid,
      neutral
    )
  )

```

```{r}

df_cleaned %>% 
  glimpse()

```

### Changing the shape of the dataset from long to wide to make it tidy

```{r}

df_cleaned <- 
  df_cleaned %>% 
  pivot_wider(
    id_cols = c(
      submission_id, 
      age,
      condition,
      gender
      ),
    names_from = trial_name,
    values_from = response
  ) %>% 
  select(
    -c(
      understanding_check
    )
  )

```

### Demographics with `NA`s dropped

```{r}

# this chunk is taken from the original study and is modified for the replication study
df_cleaned_no_na <-
  df_cleaned %>%
  drop_na()
  # select(age, gender) %>%
  
demographics_summary <- 
  tibble(mean_age = signif(mean(df_cleaned_no_na$age), 1),
            min_age = min(df_cleaned_no_na$age),
            max_age = max(df_cleaned_no_na$age),
            sd_age = signif(sd(df_cleaned_no_na$age), 1),
            median_ag = median(df_cleaned_no_na$age),
            N = length(df_cleaned_no_na$age),
            prop_males = signif(sum(df_cleaned_no_na$gender=="male")/N, 1),
            prop_females = signif(sum(df_cleaned_no_na$gender=="female")/N, 1),
            prop_rest = signif(1-(prop_males+prop_females), 1),
  )

demographics_summary

```
> NOTE: odd `max_age`


### Plotting the `age` distribution and getting their counts

```{r}

df_cleaned %>% 
  ggplot(aes(age)) +
  geom_bar()

```

```{r}

df_cleaned %>% 
  group_by(age) %>%
  count()

```

### Getting the mean of `age` without `NA`s and `age > 100`

```{r}

mean_age <- 
  df_cleaned %>% 
  filter(
    age < 100, 
    !is.na(age)
    ) %>% 
  summarise(
    value = round(mean(age))
  )

```

### Replacing the `NA`s and `age > 100` in `age` with the `mean` of the rest of ages

```{r}

df_cleaned$age <-
  if_else(
    df_cleaned$age > 100 , 
    mean_age$value, 
    df_cleaned$age, 
    missing = mean_age$value)

```


```{r}

df_cleaned %>% 
  group_by(age) %>%
  tally()

```
### Plotting `age` distribution after replacements

```{r}

df_cleaned %>% 
  ggplot(aes(age)) +
  geom_bar()

```

### Checking whether there are `NA`s in `gender`

```{r}

df_cleaned %>% 
  group_by(gender) %>%
  tally()

```
### Replacing `NA`s in `gender` with `prefer_not_to_say`

```{r}

df_cleaned$gender <-
  ifelse(is.na(df_cleaned$gender), "prefer_not_to_say", df_cleaned$gender)

```

```{r}

df_cleaned %>% 
  group_by(gender) %>%
  tally()

```

### Demographics after replacements

```{r}

demographics_summary <- 
  tibble(mean_age = signif(mean(df_cleaned$age), 1),
            min_age = min(df_cleaned$age),
            max_age = max(df_cleaned$age),
            sd_age = signif(sd(df_cleaned$age), 1),
            median_ag = median(df_cleaned$age),
            N = length(df_cleaned$age),
            prop_males = signif(sum(df_cleaned$gender=="male")/N, 1),
            prop_females = signif(sum(df_cleaned$gender=="female")/N, 1),
            prop_rest = signif(1-(prop_males+prop_females), 1) # proportion of all options that are neither female nor male
  )

demographics_summary

```

### Encoding variables for analysis according to the [`2019 Pryor et al.`](https://journals.plos.org/plosone/article?id=10.1371/journal.pone.0219464)

```{r}

df_cleaned <- 
  df_cleaned %>%
  mutate(
    ingroupNorm = case_when(
      condition == 1 |
        condition == 3 ~ -1,   # reporting robber
      condition == 2 |
        condition == 4 ~ 1    # leave him alone
    ),
    bothShown = case_when(
      condition == 1 |
        condition == 2 ~ 0,   # only ingroup
      condition == 3 |
        condition == 4 ~ 1    # both in and outgroup
      )
    )

```


### Creating variables `ingroupAgree` and `outgroupDisagree` for the alternative model

```{r}

df_cleaned <- 
  df_cleaned %>% 
  mutate(
    ingroupAgree = ifelse(
      identity_check_Pro >= 5, 1, 0   # 1 for participant identifying with the ingroup
      ),
    outgroupDisagree = ifelse(
      identity_check_Anti <= 3, 1, 0   # 1 for participant not identifying with the outgroup
      )
    ) %>% 
  rename(response = experimental_trial)

```


```{r}

df_cleaned %>%
  glimpse()

```


### Storing the cleaned dataset

```{r}

write.csv(df_cleaned,"02_clean_data.csv", row.names = FALSE)

```

# --- ` Exploration ` ---

### Counting the participants who failed the understanding checks

```{r}

df_exploration <- df %>% 
  mutate(
    invalid_U1 = case_when(
      trial_name == "understanding_check" &
        response == "incorrect" ~ TRUE,
      TRUE ~ FALSE
      ),
    invalid_U2 = case_when(
      trial_name == "understanding_check2" &
        response == "incorrect" ~ TRUE,
      TRUE ~ FALSE
      ),
    invalid_both = case_when(
      trial_name == "understanding_check" &
      trial_name == "understanding_check2" &
        response == "incorrect" ~ TRUE,
      TRUE ~ FALSE
      )
    )

invalid_ids_U1 <- df_exploration$submission_id[which(df_exploration$invalid_U1 == TRUE)]
invalid_ids_U2 <- df_exploration$submission_id[which(df_exploration$invalid_U2 == TRUE)]
invalid_ids_both <- df_exploration$submission_id[which(df_exploration$invalid_both == TRUE)]

sum_invalid_U1 <- sum(df_exploration$invalid_U1)
sum_invalid_U2 <- sum(df_exploration$invalid_U2)
sum_invalid_both <- sum(df_exploration$invalid_both)


cat(
  paste0(
    sum_invalid_U1, " participant(s) failed the first understanding check.\n",
    sum_invalid_U2, " participant(s) failed the second understanding check.\n",
    sum_invalid_both, " participant(s) failed both understanding checks."
    )
  )


```

### Storing the exploration dataset

```{r}

write.csv(df_cleaned,"03_exploratory_data.csv", row.names = FALSE)

```

## Refrence

Pryor C, Perfors A, Howe PDL (2019)
Conformity to the descriptive norms of people with
opposing political or social beliefs. PLoS ONE 14
(7): e0219464. https://doi.org/10.1371/journal.
pone.0219464
