==> analyses/02_main/02_data_plotting.Rmd <==
---
title: "Conformity of Norms Replication Plotting - main analysis"
author: "Group 5"
date: "August 25, 2021"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE, error = F, message = F, warning = F)
```

```{r libraries, include=FALSE, message=FALSE, warning=FALSE}

library(tidyverse)
library(readr)

# global color scheme / non-optimized
project_colors = c("#E69F00", "#56B4E9", "#009E73", "#F0E442", "#0072B2", "#D55E00", "#CC79A7", "#000000")

# setting theme colors globally
scale_colour_discrete <- function(...) {
  scale_colour_manual(..., values = project_colors)
}
scale_fill_discrete <- function(...) {
   scale_fill_manual(..., values = project_colors)
} 

# nicer global knitr options
knitr::opts_chunk$set(warning = FALSE, message = FALSE, 
                      cache = TRUE, fig.align = 'center')

```


### Importing the data

```{r, results='hide'}

df_to_plot <- read_csv("02_clean_data.csv")

```

### Plotting the distribution of the ratings of the chosen topic statement

```{r}
plot_rate_statement <- 
  df_to_plot %>%
  ggplot(
    aes(
      rate_statement
      ), 
    ) +
  geom_bar(
    size = 3,
    alpha = .7, 
    fill = "#1c9099"
    ) +
  labs(
    x = "Rating",
    y = "Count",
    title = "Rating the statement of the chosen topic") +
  scale_x_discrete(limits = -5:5, labels = c("Strongly \n disagree", "", "", "", "",
                                              "", "", "", "", "", "Strongly \n agree")) +
  scale_y_discrete(limits = seq(0, 14, by=2)) +
  theme_default() +
  theme(
    plot.title = element_text(size = 16, 
                              hjust = .5, 
                              margin = margin(b = 20)
                              ),
    axis.title.y = element_text(size = 14
                                , hjust = .5
                                , margin = margin(r = 15, l = 20)
                                ),
    axis.title.x = element_text(size = 14, 
                                hjust = .5,
                                margin = margin(t = 15, b = 20)
                                ),
    plot.margin = margin(t = 15, r = 35)
  )

plot_rate_statement
```

### Plotting mean response of the main trial

```{r}
plot_replica <- produce_mean_and_count_bar_plot(df_to_plot, bar_width_means=1, bar_width_response=0.3)

plot_replica
```

# Saving the plots

```{r}

ggsave(file="plot_replica_ingroupNorm-Preference.png", plot=plot_replica, width=190, height = 110, units="mm")  
ggsave(file="plot_rate_statement.png", plot=plot_rate_statement, width=190, height = 110, units="mm")  

```


```{r}
# This function is taken from the original study's OSF repository and went through minor aesthetic modifications 

produce_mean_and_count_bar_plot <- function(df, bar_width_means=0.5, bar_width_response=0.4){
  #--Find the higher response proportion so that I can rescale the other proportions relative to that. 
  response_counts = group_by(df, ingroupNorm, bothShown, response)%>%
    summarize(count = length(response))%>%ungroup()
  max_response_count <- max(response_counts)
  #--Set the dimensions of rectangles to add in to plot representing proportions of responses
  props <- group_by(df, ingroupNorm, bothShown, response) %>%
    summarize(count = length(response), 
              relative_proportion = count/max_response_count,
              xmin = ingroupNorm[1] - bar_width_means/1.5,
              xmax = xmin + relative_proportion*bar_width_means,
              ymin = response[1] - bar_width_response, 
              ymax = response[1] + bar_width_response) %>%ungroup()
  props$bothShown = factor(props$bothShown, labels=c("Only ingroup norm shown", "Both norms shown"))
  
  
  #--Find mean responses for each condition
  means <- group_by(df, ingroupNorm, bothShown) %>%
    summarize(meanResponse = mean(response),
              stdErrResponse = sd(response)/sqrt(length(response)),
              boundLower = meanResponse - stdErrResponse,
              boundUpper = meanResponse + stdErrResponse
    ) %>% ungroup()
  means$bothShown = factor(means$bothShown, labels=c("Only ingroup norm shown", "Both norms shown"))
  
  
  library(scales) #Needed to give rescale_none function to oob. 
  return_plot <- ggplot()+
    geom_col(data=means, aes(x=ingroupNorm, y=meanResponse, fill=factor(ingroupNorm)), width=bar_width_means, alpha=0.5) +
    geom_rect(data=props, aes(xmin=xmin, ymin=ymin, xmax=xmax, ymax=ymax, group=ingroupNorm), fill = "#9c9d9e", alpha = 0.8) +
    scale_y_continuous(name="Preference", limits=c(1,6), oob = rescale_none, breaks = 1:6, labels=c("Definitely \n Report", "","","","", "Definitely \n leave alone")) + 
    scale_x_continuous(name = "Ingroup Norm", breaks=c(-1,1), labels=c("Report", "Not Report"))+
    scale_fill_manual(labels=c("Report", "Not Report"), values = c("#1c9099", "#7fcdbb")) +
    facet_wrap(~bothShown) + 
    labs(title = "Replication study")+
    scale_alpha(range = c(0, 1)) +
    theme_default() +
    theme(panel.spacing = unit(4, "lines"),
          strip.text = element_text(size = 14),
          panel.spacing.x = unit(3, "cm"),
          plot.title = element_text(hjust = .5)) +
    guides(fill=FALSE, group=FALSE, alpha=FALSE)
  return(return_plot)
}
```



```{r}

df_proc <- df %>% 
  select(
    -c(
      QUD,
      RT,
      optionLeft,
      optionRight,
      startDate,
      startTime,
      trial_number,
      question
  ))

View(df_proc)
```

```{r}

df_proc <- df_proc %>% 
  pivot_wider(names_from = "trial_name", values_from = "response")

View(df_proc)

```

```{r}

df_groups <- df_proc %>% 
  count(condition)

#View(df_groups)

ggplot(df_groups, aes(x = "", y = n, fill = factor(condition))) +
    geom_bar(stat = "identity", width = 1, col = "white") +
    #scale_fill_manual("condition",values=c("yellow", "blue", "red", "green")) +
    scale_fill_brewer("condition",palette = "Set1")+
    coord_polar("y", start=0) +
    theme_void()+
    ggtitle("Count")

        
```

```{r}
df_ratings <- df_proc %>% 
  group_by(condition, rate_statement) %>% 
  count()

#(df_ratings)

ggplot(df_ratings, aes(x = as.numeric(rate_statement), y = n, fill = factor(condition) ))+
  scale_fill_brewer("condition", palette = "Set1")+
  geom_bar(stat = "identity", width = 1, col = "white")+
  scale_x_discrete(name = "rating of statement", limits = c(seq(-5,5)))+
  ylab("count")
  

```

```{r}
df_md <- df_proc %>% 
  group_by(condition, experimental_trial) %>% 
  count()

#(df_ratings)

ggplot(df_md, aes(x = as.numeric(experimental_trial), y = n, fill = factor(condition) ))+
  scale_fill_brewer("condition", palette = "Set1")+
  geom_bar(stat = "identity", width = 1, col = "white")+
  scale_x_discrete(name = "rating of statement", limits = c(seq(1,6)))+
  ylab("count")
  

```

```{r}

ggplot(df_proc, aes(x = condition, y = as.numeric(experimental_trial) , fill = factor(condition) ))+
  scale_fill_brewer("condition", palette = "Set1")+
  geom_boxplot()+
  geom_jitter(width = 0.3, height = 0.1)
 # ylim(c("report",2,3,4,5, "leave alone"))
 # scale_y_discrete(name = "moral decision", limits = c(1,2,3,4,5,6))
  #ylab("count")
  

```
The higher the rating in the experimental trial the more the participant tends to leave the robber alone. In groups 2 and 4 the participant 

```{r}
#just for fun, no use intended

df_age <- df_proc %>% 
  drop_na()

df_age <- df_age[df_age$age < 100,]

df_age["Pro_Contra"] = case_when(
                        df_age$rate_statement > 0 ~ "Pro",
                        df_age$rate_statement < 0 ~ "Contra"
                        )


#View(df_age)
ggplot(df_age, aes(x = age, 
                   y = as.numeric(experimental_trial), 
                   col = choice_of_political_topic))+
  geom_point()+
  geom_smooth(method='lm', formula= y~x, col = "black")#+
  #scale_x_discrete("Age", limits= c(0,100))
```
